// services/api/src/__generated__/graphqlgen.ts
export namespace QueryResolvers {
  export const defaultResolvers = {}

  export interface ItemFilter {
    dappAddress?: string | null
    searchString?: string | null
    inclusions?: string[] | null
  }

  export interface ArgsItems {
    filter?: ItemFilter | null
    orderBy?: ItemOrderByInput | null
    after?: string | null
    first?: number | null
  }

  export interface ArgsItem {
    assetId: string
  }

  export interface ArgsUser {
    address: string
  }

  export interface ArgsInventory {
    address: string
    orderBy?: ItemOrderByInput | null
  }
// ...
  export interface Type {
    items:
      | ((
          parent: undefined,
          args: ArgsItems,
          ctx: Context,
          info: GraphQLResolveInfo
        ) => ItemConnection | null | Promise<ItemConnection | null>)
      | {
          fragment: string
          resolve: (
            parent: undefined,
            args: ArgsItems,
            ctx: Context,
            info: GraphQLResolveInfo
          ) => ItemConnection | null | Promise<ItemConnection | null>
        }

    item:
      | ((
          parent: undefined,
          args: ArgsItem,
          ctx: Context,
          info: GraphQLResolveInfo
        ) => Item | null | Promise<Item | null>)
      | {
          fragment: string
          resolve: (
            parent: undefined,
            args: ArgsItem,
            ctx: Context,
            info: GraphQLResolveInfo
          ) => Item | null | Promise<Item | null>
        }

    dapps:
      | ((parent: undefined, args: {}, ctx: Context, info: GraphQLResolveInfo) => User[] | Promise<User[]>)
      | {
          fragment: string
          resolve: (parent: undefined, args: {}, ctx: Context, info: GraphQLResolveInfo) => User[] | Promise<User[]>
        }

    user:
      | ((
          parent: undefined,
          args: ArgsUser,
          ctx: Context,
          info: GraphQLResolveInfo
        ) => User | null | Promise<User | null>)
      | {
          fragment: string
          resolve: (
            parent: undefined,
            args: ArgsUser,
            ctx: Context,
            info: GraphQLResolveInfo
          ) => User | null | Promise<User | null>
        }

    inventory:
      | ((parent: undefined, args: ArgsInventory, ctx: Context, info: GraphQLResolveInfo) => Item[] | Promise<Item[]>)
      | {
          fragment: string
          resolve: (
            parent: undefined,
            args: ArgsInventory,
            ctx: Context,
            info: GraphQLResolveInfo
          ) => Item[] | Promise<Item[]>
        }

    me:
      | ((parent: undefined, args: {}, ctx: Context, info: GraphQLResolveInfo) => User | null | Promise<User | null>)
      | {
          fragment: string
          resolve: (
            parent: undefined,
            args: {},
            ctx: Context,
            info: GraphQLResolveInfo
          ) => User | null | Promise<User | null>
        }

    platformStats:
      | ((
          parent: undefined,
          args: {},
          ctx: Context,
          info: GraphQLResolveInfo
        ) => PlatformStats | Promise<PlatformStats>)
      | {
          fragment: string
          resolve: (
            parent: undefined,
            args: {},
            ctx: Context,
            info: GraphQLResolveInfo
          ) => PlatformStats | Promise<PlatformStats>
        }
  }
}